requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0
PyYAML>=6.0

# ML/Bayesian
jax[cpu]>=0.4.0
numpyro>=0.13.0

# RAG
chromadb>=0.4.0
sentence-transformers>=2.2.0

# LLM backends
mlx-lm>=0.10.0  # For Apple Silicon

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Utils
scipy>=1.10.0
scikit-learn>=1.3.0
statsmodels>=0.14.0

